Revert "fix a small compiler warning in libicq200"
This reverts commit 9b3cd99b2083e1af6312868d9983681230c7e0d5.

+++ libicq2000/src/FileTransferClient.cpp
@@ -201,7 +201,7 @@ namespace ICQ2000
   {
     if (m_recv.empty()) return;
 
-    unsigned int length;
+    unsigned short length;
 
     while (!m_recv.empty())
     {
